+++ lib/widgets/priority_button.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+
+import '../theme/colors.dart';
+import '../txdx/txdx_item.dart';
+
+class PriorityButton extends StatelessWidget {
+
+  final IconData iconData;
+  final String? priority;
+  final TxDxItem item;
+  final GestureTapCallback onTap;
+
+  const PriorityButton({
+    Key? key,
+    required this.iconData,
+    this.priority,
+    required this.item,
+    required this.onTap
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final selectedPriority = item.priority == priority;
+    final highlightColor = Theme.of(context).brightness == Brightness.dark
+        ? TxDxColors.darkContextHoverColor
+        : TxDxColors.lightContextHoverColor;
+
+    return Container(
+      decoration: BoxDecoration(
+        borderRadius: const BorderRadius.all(Radius.circular(8)),
+        color: selectedPriority ? highlightColor : Colors.transparent,
+      ),
+      child: IconButton(
+          icon: FaIcon(
+              iconData,
+              color: TxDxColors.forPriority(priority),
+              size: 16),
+          splashRadius: 1,
+          hoverColor: highlightColor,
+          onPressed: onTap,
+      ),
+    );
+  }
+}